Let an importer load without the plugin
ImportContext holds counts and nothing else - no vault, no DOM, no Obsidian
API - but it lived in main.ts, which imports Plugin, Modal and all fourteen
importers. Every importer took its ImportContext from there, so importing one
in a test pulled the whole plugin in with it and failed on `Class extends
value undefined`.

It moves to import-context.ts, and the three values an importer needs but the
plugin object does not carry - ATTACHMENT_EXTS, AUTH_REDIRECT_URI and
AuthCallback - to constants.ts. That also breaks the cycle between main.ts and
format-importer.ts.

Three imports used bare specifiers ('zip', 'main', 'filesystem'), which resolve
only through tsconfig.json's baseUrl of src. tsconfig.test.json sets it to '.',
so under the test runner textbundle.ts could not resolve 'zip' - and since
main.ts imports textbundle, that one line made the entire module graph
unloadable. onenote.ts, the largest importer with no tests, now loads.

ProgressReporter goes with it: the alias was marked temporary, and textbundle
was the last thing using it.
