+++ src/formats/apple-notes.ts
@@ -2,7 +2,7 @@ import { Notice, Platform, TFile, TFolder, moment } from 'obsidian';
 import { NoteConverter } from './apple-notes/convert-note';
 import { ANAccount, ANAttachment, ANContext, ANConverter, ANConverterType, ANFolderType } from './apple-notes/models';
 import { descriptor } from './apple-notes/descriptor';
-import { ImportContext } from '../main';
+import { ImportContext } from '../import-context';
 import { fsPromises, nodeBufferToArrayBuffer, os, parseFilePath, path, splitext, zlib } from '../filesystem';
 import { extractErrorMessage, sanitizeFileName } from '../util';
 import { FormatImporter } from '../format-importer';
