+++ src/formats/textbundle.ts
@@ -1,8 +1,8 @@
 import { normalizePath, Notice, TFolder, Platform } from 'obsidian';
 import { parseFilePath, NodePickedFolder, NodePickedFile, PickedFile, PickedFolder } from '../filesystem';
 import { FormatImporter } from '../format-importer';
-import { ProgressReporter } from '../main';
-import { readZip, ZipEntryFile } from 'zip';
+import { ImportContext } from '../import-context';
+import { readZip, ZipEntryFile } from '../zip';
 import { bundleNoteName, convertTextbundleNote, groupFilesByTextbundle, isMarkdownBundle } from './textbundle/convert';
 
 export class TextbundleImporter extends FormatImporter {
@@ -25,7 +25,7 @@ export class TextbundleImporter extends FormatImporter {
 		this.addOutputLocationSetting('Textbundle');
 	}
 
-	async import(progress: ProgressReporter): Promise<void> {
+	async import(progress: ImportContext): Promise<void> {
 		let { files } = this;
 		if (files.length === 0) {
 			new Notice('Please pick at least one file to import.');
@@ -62,7 +62,7 @@ export class TextbundleImporter extends FormatImporter {
 		}
 	}
 
-	async process(progress: ProgressReporter, bundleName: string, entries: (PickedFile | PickedFolder | ZipEntryFile)[]) {
+	async process(progress: ImportContext, bundleName: string, entries: (PickedFile | PickedFolder | ZipEntryFile)[]) {
 		// First look for the info.json and check that the file type is Markdown
 		const infojson = entries.find((entry) => entry.name === 'info.json');
 		if (infojson && !isMarkdownBundle(await (infojson as NodePickedFile).readText())) {
@@ -115,7 +115,7 @@ export class TextbundleImporter extends FormatImporter {
 		}
 	}
 
-	async importAsset(progress: ProgressReporter, entry: PickedFile | PickedFolder | ZipEntryFile): Promise<void> {
+	async importAsset(progress: ImportContext, entry: PickedFile | PickedFolder | ZipEntryFile): Promise<void> {
 		if (entry.type === 'folder') {
 			progress.reportSkipped(entry.name);
 			return;
